Moved tests, changes after review
